+++ src/tree/branch.rs
@@ -4,40 +4,44 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-use super::{PoseidonLeaf, PoseidonTreeAnnotation};
+use super::PoseidonLeaf;
 
-use alloc::vec::Vec;
-use canonical_derive::Canon;
+use crate::tree::PoseidonAnnotation;
+
+use core::borrow::Borrow;
 use core::iter;
 use core::ops::Deref;
+
 use dusk_bls12_381::BlsScalar;
 use dusk_hades::{ScalarStrategy, Strategy};
 use microkelvin::Branch;
+use nstack::annotation::Keyed;
 use nstack::NStack;
 
 /// Represents a level of a branch on a given depth
-#[derive(Debug, Default, Clone, Copy, Canon)]
+#[derive(Debug, Default, Clone, Copy)]
 pub struct PoseidonLevel {
     level: [BlsScalar; dusk_hades::WIDTH],
-    offset: u64,
+    index: u64,
 }
 
 impl PoseidonLevel {
     /// Represents the offset of a node for a given path produced by a branch
     /// in a merkle opening
     ///
-    /// The first position in a level set is represented as offset `1` because internally the hades
-    /// permutation preprend the bitflags for merkle opening consistency
-    pub const fn offset(&self) -> u64 {
-        self.offset
+    /// The first position in a level set is represented as offset `1` because
+    /// internally the hades permutation prepend the bitflags for merkle opening
+    /// consistency
+    pub const fn index(&self) -> u64 {
+        self.index
     }
 
     /// Represents the current level offset as a bitflag
     ///
-    /// The LSB (least significant bit) represents the offset `1`. Any increment on the offset will
-    /// left shift this flag by `1`.
+    /// The LSB (least significant bit) represents the offset `1`. Any increment
+    /// on the offset will left shift this flag by `1`.
     pub const fn offset_flag(&self) -> u64 {
-        1 << (self.offset - 1)
+        1 << (self.index - 1)
     }
 }
 
@@ -45,7 +49,7 @@ impl Deref for PoseidonLevel {
     type Target = BlsScalar;
 
     fn deref(&self) -> &Self::Target {
-        &self.level[self.offset as usize]
+        &self.level[self.index as usize]
     }
 }
 
@@ -56,7 +60,7 @@ impl AsRef<[BlsScalar]> for PoseidonLevel {
 }
 
 /// Represents a full path for a merkle opening
-#[derive(Debug, Clone, Canon)]
+#[derive(Debug, Clone)]
 pub struct PoseidonBranch<const DEPTH: usize> {
     path: Vec<PoseidonLevel>,
 }
@@ -98,13 +102,16 @@ impl<const DEPTH: usize> AsRef<[PoseidonLevel]> for PoseidonBranch<DEPTH> {
     }
 }
 
-impl<L, A, const DEPTH: usize> From<&Branch<'_, NStack<L, A>, A>>
+impl<L, K, const DEPTH: usize>
+    From<&Branch<'_, NStack<L, PoseidonAnnotation<K>>, PoseidonAnnotation<K>>>
     for PoseidonBranch<DEPTH>
 where
-    L: PoseidonLeaf,
-    A: PoseidonTreeAnnotation<L>,
+    L: PoseidonLeaf + Keyed<K>,
+    K: Clone + PartialOrd,
 {
-    fn from(b: &Branch<'_, NStack<L, A>, A>) -> Self {
+    fn from(
+        b: &Branch<'_, NStack<L, PoseidonAnnotation<K>>, PoseidonAnnotation<K>>,
+    ) -> Self {
         let mut branch = PoseidonBranch::default();
         let mut depth = 0;
 
@@ -114,7 +121,7 @@ where
             .zip(branch.path.iter_mut())
             .for_each(|(l, b)| {
                 depth += 1;
-                b.offset = l.offset() as u64 + 1;
+                b.index = l.index() as u64 + 1;
 
                 let mut flag = 1;
                 let mut mask = 0;
@@ -135,9 +142,13 @@ where
                         .iter()
                         .zip(b.level.iter_mut().skip(1))
                         .for_each(|(node, l)| {
-                            if let Some(node) = node {
+                            if let Some(annotated) = node {
+                                let anno = annotated.anno();
+                                let anno = &*anno;
+
                                 mask |= flag;
-                                *l = *node.annotation().borrow();
+
+                                *l = *anno.borrow();
                             }
 
                             flag <<= 1;
@@ -160,7 +171,7 @@ where
             perm.copy_from_slice(&l);
             h.perm(&mut perm);
 
-            b.offset = 1;
+            b.index = 1;
             b.level[0] = flag;
             b.level[1] = perm[1];