+++ CHANGELOG.md
@@ -10,6 +10,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add support for `rkyv-impl` under `no_std`
+- Add `ranno` version `0.1` to dependencies [#180]
+
+### Changed
+
+- Change `PoseidonTree` to build only with the `alloc` feature [#180]
+- Change `PoseidonTree` to take a generic `Keyed` over the leaf type [#180]
+  instead of a `PoseidonAnnotation` [#180]
+- Make `PoseidonTree::new` const [#180]
+- Update `microkelvin` from `0.15` to `0.17` [#180]
+- Update `nstack` from `0.14.0-rc` to `0.16` [#180]
+
+### Removed
+
+- Remove `std` feature [#180]
+- Remove `canon` and `persistence` features [#180]
+- Remove `Error` struct [#180]
+- Remove `canonical` and `canonical-derive` from dependencies [#180]
+- Remove `PoseidonMaxAnnotation` [#180]
 
 ## [0.26.0] - 2022-08-17
 
@@ -307,6 +325,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Poseidon252 Sponge-hash impl with BulletProofs.
 - Variants of sponge for `Scalar` & `Gadget(Variable/LC)`.
 
+[#180]: https://github.com/dusk-network/poseidon252/issues/180
 [#175]: https://github.com/dusk-network/poseidon252/issues/175
 [#167]: https://github.com/dusk-network/poseidon252/issues/167
 [#162]: https://github.com/dusk-network/poseidon252/issues/162